Refactor tests and clean up
