+++ pulseapi/entries/test_views.py
@@ -1,45 +1,11 @@
 import json
 
 from django.core.urlresolvers import reverse
-from django.test import TestCase
-from django.test import Client
 
-from pulseapi.users.test_models import EmailUserFactory
-from pulseapi.entries.test_models import EntryFactory
+from pulseapi.tests import PulseTestCase
 
 
-class TestEntryView(TestCase):
-    def setUp(self):
-        self.entries = [EntryFactory() for i in range(2)]
-        for entry in self.entries:
-            entry.save()
-
-        email = "test@example.org"
-        password = "password1234"
-        user = EmailUserFactory(email=email, password=password, name="test user")
-        user.save()
-
-        self.client = Client()
-        self.client.force_login(user);
-
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
-        # Set up with some curated data for all tests to use
-        postresponse = self.client.post('/entries/', data={
-            'title': 'title setUp1',
-            'description': 'description setUp',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
-            'creators': 'Pomax',
-            'tags': ['tag1', 'tag2'],
-            'interest': 'interest field',
-            'issues': 'Open Innovation',
-            'get_involved': 'get involved text field',
-            'get_involved_url': 'http://example.com/getinvolved',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-            'internal_notes': 'Some internal notes'
-        })
-
+class TestEntryView(PulseTestCase):
     def test_get_single_entry_data(self):
         """
         Check if we can get a single entry by its `id`
@@ -53,95 +19,46 @@ class TestEntryView(TestCase):
         """
         Test posting an entry with minimum amount of content
         """
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
-        postresponse = self.client.post('/entries/', data={
-            'title': 'test entry 1',
-            'description': 'description',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
-            'interest': 'none',
-            'get_involved': 'no',
-            'get_involved_url': 'http://example.com/',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-        })
+        payload = self.generatePostPayload(data={'title':'title test_post_minimum_entry'})
+        postresponse = self.client.post('/entries/', payload)
 
         self.assertEqual(postresponse.status_code, 200)
 
     def test_post_duplicate_title(self):
         """Make sure multiple entries can have the same title"""
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
 
-        postresponse = self.client.post('/entries/', data={
+        payload = {
             'title': 'title setUp1',
-            'description': 'description new setUp',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
-            'creators': 'Pomax',
-            'tags': ['tag1', 'tag2'],
-            'interest': 'interest field',
-            'issues': 'Open Innovation',
-            'get_involved': 'get involved text field',
-            'get_involved_url': 'http://example.com/getinvolved',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-            'internal_notes': 'Some internal notes'
-        })
+        }
+        postresponse = self.client.post('/entries/', data=self.generatePostPayload(data=payload))
         entriesJson = json.loads(str(self.client.get('/entries/').content, 'utf-8'))
         self.assertEqual(postresponse.status_code, 200)
         self.assertEqual(len(entriesJson), 4)
 
     def test_post_empty_title(self):
         """Make sure entries require a title"""
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
 
-        postresponse = self.client.post('/entries/', data={
-            'title': '',
-            'description': 'description empty title',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
-            'creators': 'Pomax',
-            'tags': ['tag1', 'tag2'],
-            'interest': 'interest field',
-            'issues': 'Open Innovation',
-            'get_involved': 'get involved text field',
-            'get_involved_url': 'http://example.com/getinvolved',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-            'internal_notes': 'Some internal notes'
-        })
+        payload = {
+            'title':''
+        }
+        postresponse = self.client.post('/entries/', data=self.generatePostPayload(data=payload))
         self.assertEqual(postresponse.status_code, 400)
 
     def test_post_empty_description(self):
         """Make sure entries require a description"""
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
 
-        postresponse = self.client.post('/entries/', data={
+        payload = {
             'title': 'title empty description',
             'description': '',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
-            'creators': 'Pomax',
-            'tags': ['tag1', 'tag2'],
-            'interest': 'interest field',
-            'issues': 'Open Innovation',
-            'get_involved': 'get involved text field',
-            'get_involved_url': 'http://example.com/getinvolved',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-            'internal_notes': 'Some internal notes'
-        })
+        }
+        postresponse = self.client.post('/entries/', data=self.generatePostPayload(data=payload))
         self.assertEqual(postresponse.status_code, 400)
 
     def test_post_full_entry(self):
         """Entry with all content"""
-
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
-        postresponse = self.client.post('/entries/', data={
+        payload = {
             'title': 'test full entry',
             'description': 'description full entry',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
             'tags': ['tag1', 'tag2'],
             'interest': 'interest field',
             'get_involved': 'get involved text field',
@@ -152,7 +69,8 @@ class TestEntryView(TestCase):
             'featured': True,
             'issues': 'Decentralization',
             'creators': ['Pomax', 'Alan']
-        })
+        }
+        postresponse = self.client.post('/entries/', data=self.generatePostPayload(data=payload))
         self.assertEqual(postresponse.status_code, 200)
 
     def test_post_entry_with_mixed_tags(self):
@@ -160,21 +78,13 @@ class TestEntryView(TestCase):
         Post entries with some existing tags, some new tags
         See if tags endpoint has proper results afterwards
         """
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
-        postresponse = self.client.post('/entries/', data={
+        payload = {
             'title': 'title test_post_entry_with_mixed_tags2',
             'description': 'description test_post_entry_with_mixed_tags',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
             'tags': ['tag2', 'tag3'],
-            'interest': 'interest field',
-            'get_involved': 'get involved text field',
-            'get_involved_url': 'http://example.com/getinvolved',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-            'internal_notes': 'Some internal notes'
-        })
-
+        }
+        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
+        postresponse = self.client.post('/entries/', data=self.generatePostPayload(data=payload))
         tagList = json.loads(str(self.client.get('/tags/').content, 'utf-8'))
         self.assertEqual(tagList, ['tag1','tag2','tag3'])
 
@@ -184,20 +94,13 @@ class TestEntryView(TestCase):
         Post entry with some existing creators, some new creators
         See if creators endpoint has proper results afterwards
         """
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
-        postresponse = self.client.post('/entries/', data={
+        payload = {
             'title': 'title test_post_entry_with_mixed_tags2',
             'description': 'description test_post_entry_with_mixed_tags',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
             'creators': ['Pomax','Alan'],
-            'interest': 'interest field',
-            'get_involved': 'get involved text field',
-            'get_involved_url': 'http://example.com/getinvolved',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-            'internal_notes': 'Some internal notes'
-        })
+        }
+        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
+        postresponse = self.client.post('/entries/', data=self.generatePostPayload(data=payload))
         creatorList = json.loads(str(self.client.get('/creators/').content, 'utf-8'))
         self.assertEqual(creatorList, ['Pomax','Alan'])
 
@@ -221,42 +124,25 @@ class TestEntryView(TestCase):
 
     def test_entries_issue(self):
         """test filtering entires by issue"""
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
-        postresponse = self.client.post('/entries/', data={
+        payload = {
             'title': 'title test_entries_issue',
             'description': 'description test_entries_issue',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
-            'creators': ['Pomax','Alan'],
-            'interest': 'interest field',
             'issues': 'Decentralization',
-            'get_involved': 'get involved text field',
-            'get_involved_url': 'http://example.com/getinvolved',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-            'internal_notes': 'Some internal notes'
-        })
+        }
+        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
+        postresponse = self.client.post('/entries/', data=self.generatePostPayload(data=payload))
         searchList = self.client.get('/entries/?issue=Decentralization')
         entriesJson = json.loads(str(searchList.content, 'utf-8'))
         self.assertEqual(len(entriesJson), 1)
 
     def test_post_entry_new_issue(self):
         """posting an entry with a new Issue should result in an error. Permission denied?"""
-        values = json.loads(str(self.client.get('/nonce/').content, 'utf-8'))
-        postresponse = self.client.post('/entries/', data={
+        payload = {
             'title': 'title test_entries_issue',
             'description': 'description test_entries_issue',
-            'nonce': values['nonce'],
-            'csrfmiddlewaretoken': values['csrf_token'],
-            'creators': ['Pomax','Alan'],
-            'interest': 'interest field',
             'issues': 'Privacy',
-            'get_involved': 'get involved text field',
-            'get_involved_url': 'http://example.com/getinvolved',
-            'thumbnail_url': 'http://example.com/',
-            'content_url': 'http://example.com/',
-            'internal_notes': 'Some internal notes'
-        })
+        }
+        postresponse = self.client.post('/entries/', data=self.generatePostPayload(data=payload))
         self.assertEqual(postresponse.status_code, 400)
 
     def test_post_authentication_requirement(self):
